requirements.txt
streamlit
google-generativeai
pypdf
numpy
python-dotenv
docx2txt
